update unit test
